HW1 attatchment
These are Jpegs because I could not combine the pdfs. I got this to work at 12:29am. This portion of the assignment was completed before 9:48 pm as indicated by the changes I made on the google docs as that was after I had done 1-4. I apologize for the scattered documents but I cannot get them to bundle together in one document and none of my electronics are cooperating in a timely manner. I apologize once again and will see you in class.