README.md
I had a lot of tech issues and problems with github. I irresponsibly was doing the programming part of the assingment from 5pm until the deadline and this resulted in time issues where I published what was working 4 times (and the fourth was stamped after 12am) and when I finally got the rest to work it was at latest 12:36 am. I will clean up my files but I would like help as I do not understand how to work this as well as I thought and could not attatch HW1 to HW in my populationecology folder and so forth. Help is much needed and appreciated.